requirements.txt
Django==1.5.1
MySQL-python==1.2.4
django-tastypie==0.9.15
mimeparse==0.1.3
numpy==1.7.1
psycopg2==2.5
pyproj==1.9.3
python-dateutil==2.1
pytz==2013b
scipy==0.12.0
six==1.3.0
wsgiref==0.1.2


